task3/flight_service/requirements.txt
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
redis==5.0.4
